+++ MLEM.Ui/Elements/Panel.cs
@@ -323,6 +323,7 @@ namespace MLEM.Ui.Elements {
         /// </summary>
         protected virtual void ScrollSetup() {
             this.childrenDirtyForScroll = false;
+            this.EnsureSufficientExecutionStack();
 
             if (!this.scrollOverflow || this.IsHidden)
                 return;