+++ CHANGELOG.md
@@ -3,7 +3,9 @@ MLEM uses [semantic versioning](https://semver.org/). Potentially breaking chang
 
 ## 8.0.1 (In Development)
 
-No code changes
+### MLEM.Ui
+Improvements
+- Elements now check for a sufficient execution stack during recursive operations as a fail-safe for bugs in the layouter
 
 ## 8.0.0